refact: adding src folder and moving each file to its own folder
