--- api/.env.test
@@ -1,4 +1,0 @@
-DB_URL=
-PORT=
-DB_NAME=
-DB_PASSWD=
