--- api/routes/index.js
@@ -1,9 +1,0 @@
-const accountRoutes = require('./users');
-const loginRoutes = require('./login');
-
-module.exports = function (app) {
-  app.use('/api', [
-    accountRoutes,
-    loginRoutes
-  ])
-};
