+++ src/.env.test
@@ -1,0 +1,8 @@
+PORT=3003
+DB_USER=postgres
+DB_PWD=postgres
+DB_NAMES=ser_fit
+DB_HOST=host.docker.internal
+DB_PORT=5432
+DB_DIALECT=postgres
+SECRET=mysecret
