+++ src/__tests__/test.spec.js
@@ -1,0 +1,8 @@
+const db = require('../config/database');
+
+describe('env', () => {
+  it('db should be defined', () => {
+    console.log('db', db);
+    expect(db).toBeDefined();
+  });
+});
