+++ src/api/login/login.routes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const LoginController = require('../controllers/login');
+const LoginController = require('./login');
 const PATH = "/login";
 
 router.post(PATH + "/", LoginController.login);
