+++ src/api/login/login.test.js
@@ -1,0 +1,13 @@
+const { login } = require('./login');
+
+describe('login', () => {
+    it('should return a token', async () => {
+        const req = {
+            body: {
+                email: ''
+            }}
+        })
+});
+test('login', () => {
+    expect(login).toBeDefined();
+});
