+++ src/jestSetup.js
@@ -1,0 +1,1 @@
+process.env.ENVIROMENT = 'test';
