+++ src/package.json
@@ -22,14 +22,7 @@
     "nodemon": "^2.0.19",
     "pg": "^8.7.3",
     "sequelize": "^6.21.3",
-    "sqlite3": "^5.0.10",
     "supertest": "^6.2.4"
-  },
-  "jest": {
-    "testEnvironment": "node",
-    "coveragePathIgnorePatterns": [
-      "/node_modules/"
-    ]
   },
   "devDependencies": {
     "eslint": "^8.19.0",
