+++ src/routes/index.js
@@ -1,0 +1,9 @@
+const accountRoutes = require('../api/user/user.routes');
+const loginRoutes = require('../api/login/login.routes');
+
+module.exports = function (app) {
+  app.use('/api', [
+    accountRoutes,
+    loginRoutes
+  ])
+};
